# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QTXpert.ai Test Design Agent",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/generate-testcases": {
      "post": {
        "tags": [
          "test-cases"
        ],
        "summary": "Generate Testcases",
        "description": "Runs the full Step 1-11 AI workflow and persists the generated test cases.",
        "operationId": "generate_testcases_api_v1_generate_testcases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateTestCasesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationRunOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/history": {
      "get": {
        "tags": [
          "test-cases"
        ],
        "summary": "History",
        "operationId": "history_api_v1_history_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GenerationRunOut"
                  },
                  "title": "Response History Api V1 History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/history/{run_id}": {
      "get": {
        "tags": [
          "test-cases"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_v1_history__run_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationRunOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenerateTestCasesRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "requirement_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Requirement Ids",
            "description": "If empty, all requirements in the project are used."
          },
          "llm_provider_override": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Provider Override"
          }
        },
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "GenerateTestCasesRequest"
      },
      "GenerationRunOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "llm_provider": {
            "type": "string",
            "title": "Llm Provider"
          },
          "llm_model": {
            "type": "string",
            "title": "Llm Model"
          },
          "requirement_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requirement Summary"
          },
          "business_rules": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Rules"
          },
          "functional_breakdown": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Functional Breakdown"
          },
          "test_scenarios": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Scenarios"
          },
          "risk_analysis": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Analysis"
          },
          "processing_time_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Time Seconds"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "test_cases": {
            "items": {
              "$ref": "#/components/schemas/TestCaseOut"
            },
            "type": "array",
            "title": "Test Cases"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "status",
          "llm_provider",
          "llm_model",
          "requirement_summary",
          "business_rules",
          "functional_breakdown",
          "test_scenarios",
          "risk_analysis",
          "processing_time_seconds",
          "error_message",
          "created_at"
        ],
        "title": "GenerationRunOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Priority": {
        "type": "string",
        "enum": [
          "critical",
          "high",
          "medium",
          "low"
        ],
        "title": "Priority"
      },
...
      "RiskLevel": {
        "type": "string",
        "enum": [
          "high",
          "medium",
          "low"
        ],
        "title": "RiskLevel"
      },
      "RunStatus": {
        "type": "string",
        "enum": [
          "pending",
          "normalizing",
          "analyzing",
          "generating_scenarios",
          "generating_test_cases",
          "risk_analysis",
          "completed",
          "failed"
        ],
        "title": "RunStatus"
      },
      "Severity": {
        "type": "string",
        "enum": [
          "blocker",
          "critical",
          "major",
          "minor",
          "trivial"
        ],
        "title": "Severity"
      },
      "TestCaseOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "test_case_key": {
            "type": "string",
            "title": "Test Case Key"
          },
          "requirement_traceability": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requirement Traceability"
          },
          "test_type": {
            "$ref": "#/components/schemas/TestCaseType"
          },
          "scenario": {
            "type": "string",
            "title": "Scenario"
          },
          "objective": {
            "type": "string",
            "title": "Objective"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "preconditions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preconditions"
          },
          "test_data": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Data"
          },
          "steps": {
            "items": {},
            "type": "array",
            "title": "Steps"
          },
          "expected_result": {
            "type": "string",
            "title": "Expected Result"
          },
          "post_conditions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Post Conditions"
          },
          "is_automation_candidate": {
            "type": "boolean",
            "title": "Is Automation Candidate"
          },
          "automation_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Automation Type"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          }
        },
        "type": "object",
        "required": [
          "id",
          "test_case_key",
          "requirement_traceability",
          "test_type",
          "scenario",
          "objective",
          "priority",
          "severity",
          "preconditions",
          "test_data",
          "steps",
          "expected_result",
          "post_conditions",
          "is_automation_candidate",
          "automation_type",
          "risk_level"
        ],
        "title": "TestCaseOut"
      },
      "TestCaseType": {
        "type": "string",
        "enum": [
          "functional",
          "negative",
          "boundary",
          "integration",
          "api",
          "database",
          "security",
          "accessibility",
          "usability",
          "performance",
          "regression",
          "exploratory",
          "data_validation",
          "workflow",
          "role_based",
          "permission",
          "maker_checker",
          "error_handling",
          "localization",
          "browser_compatibility",
          "mobile",
          "cross_platform"
        ],
        "title": "TestCaseType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}